Introducing nmea module from PFWP with dependecies, in this case:
 - libswiftnav-common: provides commond gnss structures and utilities
 - libswiftnav/observation.h: modified to provide constant defines
 - PFWP/sbp_utils.h: some utilities used for sbp manipulation

These dependencies are subject to change depending on the depth of
refactor to be performed in the nmea generation APIs. The goal of
this commit was to show that the code can be compiled and exercised
outside of PFWP.

The one test (GGA) for now is not functional. Tests will be added
individually for each case to be validated in code review.